requirements.txt
streamlit
pandas
bcrypt
openpyxl
plotly
mysql-connector-python
